Run Pylint in "errors only" mode
This changes .pylintrc-default to use the configuration for errors
only mode. This results in fewer (and more significant) messages from
Pylint.

I think these messages are likely to be more useful than the current
configuration, and it will also be easier to adapt software to make
use of Pylint this way.