WIP: Make Pylint more flexible with a YAML-based config
See also: pytools#4

This adds support for a YAML-based config file format for Pylint with
a thin wrapper script that translates the YAML file into command line
arguments.

This change also modifies prepare-and-run-pylint.sh to look for YAML
configs. It looks for a .pylintrc.yml, and if it does not find it, it
downloads the global config file from ci-support. It also supports
local configurations via the file .pylintrc-local.yml.